Task 2c Step 5: Final changes
I updated the directories in Management.do, Analysis.do and the master file. I updated the Projectmap_AK91 and I deleted the old master.log and some old output results that I don't think it is useful to have here. Then I update the .gitignore to not have .tex and .log files.